backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic[email]==2.9.2
PyJWT==2.9.0
bcrypt==4.2.0
